Refactor alpine modules
